(dev/drupal#52; REF) Use more conventional polymorphism
There are ~15 helpers in `CRM_Utils_System` that are just pass-throughs.
These follow a pattern of (a) declaring a docblock, (b) delegating the
function-call via `__callStatic()`, and (c) putting the code in
`CRM_Utils_System_{Base,Drupal,WordPress,etc}` using class-method
polymorphism.

The call to `if (method_exists(...))` is essentially a similar
passthrough/class-method polymorphism.

This just changes the notation to match the others.  For someone auditing
the Civi<=>UF contract, this should make it easier to reason about the code.